% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_access.R
\name{baadclimate_access_function}
\alias{baadclimate_access_function}
\title{Baadclimate Access}
\usage{
baadclimate_access_function(version = NULL, path = NULL)
}
\arguments{
\item{version}{Version number.  The default will load the most
recent version on your computer or the most recent version known
to the package if you have never downloaded the data before.
With \code{plant_lookup_del}, specifying \code{version=NULL}
will delete \emph{all} data sets.}

\item{path}{Path to store the data at.  If not given,
\code{datastorr} will use \code{rappdirs} to find the best place
to put persistent application data on your system.  You can
delete the persistent data at any time by running
\code{mydata_del(NULL)} (or \code{mydata_del(NULL, path)} if you
use a different path).}
}
\description{
This function provides access to the baadclimate versioned dataset
}
\examples{
#
# see the format of the resource
#
#
#
}
